Move definitions from header file to cpp file
Clang-tidy does not like definitions in header files unless they're
inline.
